fix: Relocate AtpBlueprintMapping class to comply with CODING_RULE_STYLE_00008 and CODING_RULE_STYLE_00009
- Moved AtpBlueprintMapping from GenericStructure/AbstractStructure.py to CommonStructure/StandardizationTemplate/AbstractBlueprintStructure/AtpBlueprint.py
- Updated __init__.py to export AtpBlueprintMapping from correct module
- AtpBlueprintMapping now importable from correct location per mapping.json

Class mapping progress:
- Before: 733 correct (37.9%), 51 wrong
- After: 734 correct (37.9%), 50 wrong
- Progress: 4 of 54 violations fixed (7.4%)
