+++ reports/run_tests_report.md
@@ -12,7 +12,7 @@
 
 | Metric | Covered | Valid | Percentage |
 |--------|---------|-------|------------|
-| **Lines** | 23620 | 30365 | **77.79%** |
+| **Lines** | 23621 | 30366 | **77.79%** |
 | **Branches** | 0 | 0 | **0.00%** |
 
 ### Coverage by Module
