+++ src/armodel/models/M2/AUTOSARTemplates/CommonStructure/StandardizationTemplate/AbstractBlueprintStructure/__init__.py
@@ -2,6 +2,6 @@
 AbstractBlueprintStructure module for AUTOSAR M2 models.
 """
 
-from armodel.models.M2.AUTOSARTemplates.CommonStructure.StandardizationTemplate.AbstractBlueprintStructure.AtpBlueprint import AtpBlueprint
+from armodel.models.M2.AUTOSARTemplates.CommonStructure.StandardizationTemplate.AbstractBlueprintStructure.AtpBlueprint import AtpBlueprint, AtpBlueprintMapping
 
-__all__ = ['AtpBlueprint']
+__all__ = ['AtpBlueprint', 'AtpBlueprintMapping']
